Revert "Add WithResolverUserOptions for custom resolver build options (#1711)"
This reverts commit ff1be3fcc57773d500921fad479d82ec171e2358.
